Refactor parseLicenceBuffer
