+++ src/getLicenceInfo.js
@@ -3,6 +3,7 @@
 import fs from 'fs';
 import path from 'path';
 import os from 'os';
+import { unconvoluteBuffer } from './util/unconvoluteBuffer';
 
 
 // We haven't figured out exactly how these AB1E files work, so this parser is
@@ -15,15 +16,13 @@ import os from 'os';
 
 async function parseLicenceBuffer(index, buf){
 
-	let licence = {
-	};
+	let licence = {};
 
 	// Position in file:
 	licence.logicalId = index;
 
 	// read encoded index, uint16 directly after LICENCE:
 	licence.licenceId = buf.readInt16LE(0);
-
 
 	// Read product type:
 	licence.productIdRaw = Buffer.from([buf[29], buf[28]]);
@@ -38,35 +37,9 @@ async function parseLicenceBuffer(index, buf){
 	}
 
 	// Read serial number:
-	const f = buf.slice(4,26);
+	const serialBytes = unconvoluteBuffer(buf.slice(4,26));
 
-	const temp = [];
-	const sn = [];
-
-	// Holy smokes, I'm either stupid or insufficiently caffinated, I feel like
-	// I'm code golfing trying to shuffle some bytes around.  
-
-	// Create a new array with every 2nd pair of chars removed; these are nulls,
-	// but valid data can be 0x00 so we need to go by position.
-	// AABB0000CCDD0000 -> AABBCCDD
-	for (let i in [...f]) {
-		if ((!((i - 3) % 4) || (!((i - 2) % 4)))) continue;
-		temp.push(f[i]);
-	}
-
-	// Swap pairs, AABBCCDD -> BBAADDCC
-	for (let [i, n] of temp.entries()) {
-		if (i % 2) { // It doesn't work without implicit bool cast...
-			sn[i-1] = temp[i];
-		} else {
-			sn[i+1] = temp[i];
-		}
-	}
-
-	const serialBytes = Buffer.from(sn);
-
-	licence.serial = serialBytes
-		.toString('hex')
+	licence.serial = serialBytes.toString('hex')
 		.match(/.{4}/g)
 		.join('-')
 		.toUpperCase();
@@ -75,9 +48,8 @@ async function parseLicenceBuffer(index, buf){
 	licence.serialBuffer = serialBytes;
 
 	// Read the "DistrobutionType"
-	// We can infer from the header structure 
-	let dm = buf[40]; //buf.slice(40,41);
-	licence.distrobutionType = dm;
+	// We can infer name of this field from the header structure 
+	licence.distrobutionType = buf[40];
 
 	// Read responce code:
 	const rc = [];
